+++ tests/test_polars_chain.py
@@ -12,7 +12,7 @@ df = pl.DataFrame(
         "c": [3, 0, 3],
         "n": [4, 0, 4],
         "pi": [np.pi] * 3,
-        "p": [2, 0.2, 0.05],
+        "p": [2.0, 0.2, 0.05],
         "word": ["dog_cat", "dog_cat_mouse", "dog"],
     },
 )