+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.4.8] - Unreleased
 ### Added
-* n/a
+* Support for referencing nodes in SQL queries
 ### Fixed
 * n/a
 ### Updated